move new map editor to actual page
